requirements.txt
pygame>=2.5.0
Flask>=3.0.0
requests>=2.31.0
RPi.GPIO>=0.7.1;platform_machine=='armv7l' or platform_machine=='aarch64'
